Update Streamlit_app.py
Created a function called insert_row_snowflake, that will add a new fruit to the table in snowflake. The fruit is determined by whatever the end user enters on the page and then clicks the button. Finally a thank you message is returned! Only issue right now is that 'from streamlit' is still hard coded in.